backend/requirements.txt
flask
flask-cors
gunicorn
pdf2docx
python-docx
reportlab
simplekml
pandas
openpyxl